wc : modif final

+++ final/library/src/Queue.h
@@ -3,7 +3,7 @@
  * Name: Queue.h
  * Description: Random access circular buffer data structure
  * 
- * Version: 1.1
+ * Version: 1.0
 \******************************************************************************/
 
 #ifndef QUEUE_H
@@ -26,19 +26,19 @@ public:
 	 * \brief add an element to the queue, overwriting the oldest element
 	 * \param data the data to add
 	 */
-	void push(uint16_t data);
+	void push(uint8_t data);
 	
 	/**
 	 * \brief gives the element at index i, where 0 is the most recent element
 	 * \param i index
 	 * \return the element at index i
 	 */
-	uint16_t get(uint8_t i) const;
+	uint8_t get(uint8_t i) const;
 
 private:
 	
 	// Array to contain the data
-	uint16_t data_[QUEUE_CAPACITY];
+	uint8_t data_[QUEUE_CAPACITY];
 	
 	// Index where the next value will be written
 	uint8_t head_;
